+++ sources/Colonization/AngelScriptBinders/BindUnit.cpp
@@ -1,8 +1,8 @@
 #include <Colonization/BuildConfiguration.hpp>
-#include "BindUnits.hpp"
+#include "BindUnit.hpp"
 #include <Urho3D/ThirdParty/AngelScript/angelscript.h>
 #include <Urho3D/AngelScript/APITemplates.h>
-#include <Colonization/Core/Units.hpp>
+#include <Colonization/Core/Unit.hpp>
 #include <Colonization/AngelScriptBinders/BindingMacroses.hpp>
 
 namespace Colonization
